Clean up e2e tests file (#383)
* separate e2e tests by category

* Remove unused deps

* Update for v0-reg account mergo

* Fix bad merge

* More delineation

* Remove unused imports

* More delineation

* Remove copied tests

* Remove unused imports
