+++ full-service/src/json_rpc/e2e_tests/account/mod.rs
@@ -1,0 +1,4 @@
+mod account_address;
+mod account_balance;
+mod account_other;
+mod create_import;
