+++ full-service/src/json_rpc/e2e_tests/mod.rs
@@ -1,0 +1,4 @@
+mod account;
+mod gift_codes;
+mod other;
+mod transaction;
