+++ full-service/src/json_rpc/e2e_tests/transaction/build_submit/mod.rs
@@ -1,0 +1,4 @@
+mod build_and_submit;
+mod build_then_submit;
+mod large_transaction;
+mod multiple_outlay;
