+++ full-service/src/json_rpc/e2e_tests/transaction/mod.rs
@@ -1,0 +1,3 @@
+mod build_submit;
+mod transaction_other;
+mod transaction_txo;
